# src/main/java/com/challenge/guestparcelservice/representation/GuestCheckOutRequest.java
package com.challenge.guestparcelservice.representation;

import com.fasterxml.jackson.annotation.JsonProperty;

import javax.validation.constraints.NotNull;

public class GuestCheckOutRequest {

    @JsonProperty("GuestId")
    @NotNull
    private String guestId;

    public String getGuestId() {
        return guestId;
    }

    public void setGuestId(String guestId) {
        this.guestId = guestId;
    }

}