# src/main/resources/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Guest Parcel Service API Specification",
    "description": "Swagger for Guest Parcel Service API Specification",
    "contact": {
      "name": "Venkata Raghu Vamsi Chimalamarri",
      "email": "rchimalamarri@gmail.com"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/hotel/guest/check/in": {
      "post": {
        "tags": [
          "Guest Check In"
        ],
        "summary": "Create Guest Check In",
        "requestBody": {
          "description": "Default",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/GuestCheckOutResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/hotel/guest/check/out": {
      "put": {
        "tags": [
          "Guest Check Out"
        ],
        "summary": "OK",
        "requestBody": {
          "description": "Default",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestCheckOutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Guest check out details updated",
            "headers": {
              "x-fapi-interaction-id": {
                "description": "An RFC4122 UID used as a correlation id.",
                "style": "simple",
                "explode": false,
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/GuestCheckOutResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "405": {
            "description": "Method Not Allowed"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/hotel/guest/info/{date}": {
      "get": {
        "tags": [
          "GuestDetails"
        ],
        "summary": "Get GuestDetails",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "date",
            "required": true,
            "schema": {
              "pattern": "yyyy-MM-dd",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/GuestInfoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "405": {
            "description": "Method Not Allowed"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/hotel/guest/parcel/accept": {
      "put": {
        "tags": [
          "Guest Parcel  Accept"
        ],
        "summary": "Create Guest Parcel  Accept",
        "requestBody": {
          "description": "Default",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParcelAcceptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "x-fapi-interaction-id": {
                "description": "An RFC4122 UID used as a correlation id.",
                "style": "simple",
                "explode": false,
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ParcelAcceptResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GuestDetails": {
        "required": [
          "FirstName",
          "LastName",
          "ParcelCollectionEligibility"
        ],
        "type": "object",
        "properties": {
          "GuestId": {
            "type": "string",
            "description": "UUID unique identifier "
          },
          "FirstName": {
            "type": "string",
            "description": "Guest First Name"
          },
          "LastName": {
            "type": "string",
            "description": "Guest Last Name"
          },
          "CheckInDateAndTime": {
            "type": "string",
            "description": "Guest check in Date time",
            "format": "date-time"
          },
          "ParcelCollectionEligibility": {
            "type": "boolean",
            "description": "true or false"
          }
        }
      },
      "ParcelAcceptRequest": {
        "required": [
          "GuestId"
        ],
        "type": "object",
        "properties": {
          "GuestId": {
            "type": "string",
            "description": "UUID unique identifier "
          },
          "CollectedDate": {
            "type": "string",
            "description": "Guest check in Date time",
            "format": "date-time"
          },
          "AcceptedDate": {
            "type": "string",
            "description": "Guest  parcel accepted Date",
            "format": "date-time"
          },
          "IsParcelAccepted": {
            "type": "boolean",
            "description": "Guest Parcel Accepted"
          },
          "isParcelCollected": {
            "type": "boolean",
            "description": "Guest ParcelCollected"
          }
        }
      },
      "GuestCheckOutRequest": {
        "required": [
          "GuestId"
        ],
        "type": "object",
        "properties": {
          "GuestId": {
            "type": "string",
            "description": "UUID unique identifier "
          }
        }
      }
    },
    "responses": {
      "GuestInfoResponse": {
        "description": "Guest Info details for the date",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GuestDetails"
            }
          }
        }
      },
      "GuestCheckInResponse": {
        "description": "Guest Check In Response ",
        "content": {
          "application/json": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "GuestCheckOutResponse": {
        "description": "Guest Check Out Response ",
        "content": {
          "application/json": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "ParcelAcceptResponse": {
        "description": "Guest Parcel Accept Response ",
        "content": {
          "application/json": {
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}